+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.6.0]
+
+* Expose currentStepIndex on intro instance.
+* Fix bug.
+
+
+## [0.5.1]
+
+* Fix bug
+
+## [0.5.0]
+
+* Add dispose method to destroy the guide page
+
 ## [0.4.1]
 
 * Fix style issues that may occur in some cases